Update patch set 3
Patch Set 3:

(6 comments)

I will revert to Copyright 2016, The CyanogenMod Open Source Project
How about LED ON/OFF intervals?
does it make any trouble?
Please before criticize the code test it
so after that with negotiating we can make the device better.
this codes are tested on Kenzo and they are happy with the phone.

Patch-set: 3
Label: Verified=0